Modified the npc

+++ npc/csrc/sdb/sdb.h
@@ -1,12 +1,12 @@
 #ifndef __SDB_SDB_H__
 #define	__SDB_SDB_H__
 
-#include<stdio.h>
-#include<stdint.h>
-#include"../include/utils.h"
-#include"../include/debug.h"
-#include"../memory/pmem.h"
-#include"../cpu/cpu.h"
+#include <stdio.h>
+#include <stdint.h>
+#include <utils.h>
+#include <debug.h>
+#include <pmem.h>
+#include <cpu.h>
 
 #define MAX_STR_LEN 100
 #define MAX_CMD_LEN 50
@@ -18,8 +18,8 @@ void sdb_main();
 int cmd_hello(char args[]);
 
 void exec_once();
-void execute(uint32_t n);
-void cpu_exec(uint32_t n);
+void execute(uint64_t n);
+void cpu_exec(uint64_t n);
 
 void isa_reg_display();
